:boom: Version 1.0.0

+++ test/tests.js
@@ -1,10 +1,8 @@
 'use strict';
 let assert = require('chai').assert;
-const exec = require('child_process').exec;
-const execSync = require('child_process').execSync;
 const spawn = require('child_process').spawn;
 const fs = require('fs');
-const path = require("path");
+const path = require('path');
 
 const pathSep = path.sep;
 
@@ -16,10 +14,10 @@ let folder = 'testproject3';
  * @param {string} path
  */
 let deleteFolderRecursive = function(path) {
-    if( fs.existsSync(path) ) {
-        fs.readdirSync(path).forEach(function(file,index){
+    if ( fs.existsSync(path) ) {
+        fs.readdirSync(path).forEach(function(file, index) {
             let curPath = path + pathSep + file;
-            if(fs.lstatSync(curPath).isDirectory()) { // recurse
+            if (fs.lstatSync(curPath).isDirectory()) { // recurse
                 deleteFolderRecursive(curPath);
             } else { // delete file
                 fs.unlinkSync(curPath);
@@ -29,66 +27,66 @@ let deleteFolderRecursive = function(path) {
     }
 };
 
+before(function(done) {
+    this.timeout(5000);
+    deleteFolderRecursive(folder);
+    done();
+});
 
 describe('new <project>', function() {
-
     it('should create a project', function(done) {
+        this.timeout(50000);
 
-        this.timeout(40000);
-
-        exec('node jovo.js new '+folder,
-            (error, stdout, stderr) => {
-                if(error) {
-                    console.log(error);
-                    return;
-                }
-                assert.ok(stdout.indexOf('You\'re all set.') > -1);
+        let child = spawn('node', ['index.js', 'new', folder, '-t', 'helloworldtest'], {
+        });
+        child.stdout.on('data', (data) => {
+            if (data.indexOf('Installation completed.') > -1) {
+                child.kill();
                 done();
-            });
+            }
+        });
     });
 
     it('should start the webhook without errors', function(done) {
         this.timeout(10000);
-        let child = spawn('node', ['index.js'], {
+        let child = spawn('node', ['index.js', '--webhook'], {
             cwd: folder,
             // detached: true,
         });
         child.stdout.on('data', (data) => {
-            assert.ok(data.indexOf('Local development server listening on port 3000.') > -1);
-            assert.ok(data.indexOf('error') === -1);
-            child.kill();
-            done();
+            if (data.indexOf('error') > -1) {
+                assert.ok(false);
+            }
+            if (data.indexOf('Example server listening on port 3000!') > -1) {
+                child.kill();
+                done();
+            }
         });
-
     });
-
 });
 
-//TODO: more tests required
+// TODO: more tests required
 describe('bst-proxy', function() {
     it('should start the webhook without errors', function(done) {
         this.timeout(15000);
-        let child = spawn('node', ['..'+path.sep+'jovo.js', 'run', 'index.js', '--bst-proxy'], {
+        let child = spawn('node', ['..'+path.sep+'index.js', 'run', 'index.js', '--bst-proxy'], {
             cwd: folder,
-            // detached: true,
         });
         let fullData = '';
         child.stdout.on('data', (data) => {
-            fullData += data.toString();
+            fullData += data.toString('utf8');
         });
         setTimeout(() => {
             const validation =
                 // If proxy has already being run a configuration exists
-                fullData.indexOf('Local development server listening on port 3000.') > -1 ||
+                fullData.indexOf('Example server listening on port 3000!') > -1 ||
                 // If proxy haven't run, one is created
                 fullData.indexOf('info: CONFIG      No configuration. Creating one') > -1;
             assert.ok(validation);
             child.kill();
             done();
-        }, 3000)
-
+        }, 3000);
     });
-
 });
 
 after(function(done) {
